+++ addon.setup.php
@@ -12,6 +12,7 @@ return array(
     'settings_exist' => false,
     'filesystem_adapters' => [
         \CloudFiles\Adapter\AwsS3::class,
-        \CloudFiles\Adapter\DigitalOcean::class
+        \CloudFiles\Adapter\DigitalOcean::class,
+        \CloudFiles\Adapter\CloudflareR2::class,
     ]
 );