+++ language/english/cloud_files_lang.php
@@ -2,4 +2,5 @@
 $lang = [
     'adapter_awss3' => 'AWS S3',
     'adapter_digitalocean' => 'DigitalOcean Spaces',
+    'adapter_cloudflarer2' => 'Cloudflare R2',
 ];